refactor: hardhat config
