+++ deploy/01_mastercopy_module.ts
@@ -4,7 +4,7 @@ import { HardhatRuntimeEnvironment } from "hardhat/types"
 import { deployFactories, deployMastercopy } from "@gnosis-guild/zodiac-core"
 
 import createAdapter from "./eip1193"
-import MODULE_CONTRACT_ARTIFACT from "../build/artifacts/contracts/MyModule.sol/MyModule.json"
+import MODULE_CONTRACT_ARTIFACT from "../artifacts/contracts/MyModule.sol/MyModule.json"
 
 const FirstAddress = "0x0000000000000000000000000000000000000001"
